Upgrade status-go to improve indexing performance.
We were comparing the `hide` field with inequality, and this caused some
performance issues as we could not leverage the index, resulting in a
table scan.
The behavior is that if > 20 messages were present in the chat, it would
return very quickly, as the table scan would be interrupted.
If a chat had less than 20 messages, a full table scan was required.

https://github.com/status-im/status-go/pull/1986